fix: component naming
